+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - 決済カード認証タスクを追加
+- 決済URL発行タスクを追加
 
 ### Changed