+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 決済URL発行タスクを追加
 - define acceptPayAction
 - support async accpetCOAOffer
-- export IMkknInfo,IYkknInfo on checkMovieTicketAction
+- export IMkknInfo,IYkknInfo,IPurchaseNumberInfo on checkMovieTicketAction
 
 ### Changed